Refactor the code for building the entity proto
